refactor: move views write tests from `tests/queries` to `tests/writes`
