+++ query-engine/connector-test-kit-rs/query-engine-tests/tests/queries/views.rs
@@ -73,85 +73,6 @@ mod views {
         Ok(())
     }
 
-    #[connector_test]
-    async fn no_create_one_mutation(runner: Runner) -> TestResult<()> {
-        test_no_toplevel_mutation(
-            runner,
-            "no_create_one_mutation",
-            r#"mutation { createOneTestView(data: { firstName: "Test", lastName: "User", fullName: "Test User" }) { id } }"#
-        ).await
-    }
-
-    #[connector_test]
-    async fn no_update_one_mutation(runner: Runner) -> TestResult<()> {
-        test_no_toplevel_mutation(
-            runner,
-            "no_update_one_mutation",
-            r#"mutation { updateOneTestView(where: { id: 1 }, data: { firstName: "Updated" }) { id } }"#,
-        )
-        .await
-    }
-
-    #[connector_test]
-    async fn no_delete_one_mutation(runner: Runner) -> TestResult<()> {
-        test_no_toplevel_mutation(
-            runner,
-            "no_delete_one_mutation",
-            r#"mutation { deleteOneTestView(where: { id: 1 }) { id } }"#,
-        )
-        .await
-    }
-
-    #[connector_test]
-    async fn no_upsert_one_mutation(runner: Runner) -> TestResult<()> {
-        test_no_toplevel_mutation(
-            runner,
-            "no_upsert_one_mutation",
-            r#"mutation { upsertOneTestView(where: { id: 1 }, create: { firstName: "New", lastName: "User", fullName: "New User" }, update: { firstName: "Updated" }) { id } }"#
-        ).await
-    }
-
-    #[connector_test]
-    async fn no_create_many_mutation(runner: Runner) -> TestResult<()> {
-        test_no_toplevel_mutation(
-            runner,
-            "no_create_many_mutation",
-            r#"mutation { createManyTestView(data: [{ firstName: "Test", lastName: "User", fullName: "Test User" }]) { count } }"#
-        ).await
-    }
-
-    #[connector_test]
-    async fn no_update_many_mutation(runner: Runner) -> TestResult<()> {
-        test_no_toplevel_mutation(
-            runner,
-            "no_update_many_mutation",
-            r#"mutation { updateManyTestView(where: { id: 1 }, data: { firstName: "Updated" }) { count } }"#,
-        )
-        .await
-    }
-
-    #[connector_test]
-    async fn no_delete_many_mutation(runner: Runner) -> TestResult<()> {
-        test_no_toplevel_mutation(
-            runner,
-            "no_delete_many_mutation",
-            r#"mutation { deleteManyTestView(where: { id: { gt: 0 } }) { count } }"#,
-        )
-        .await
-    }
-
-    async fn test_no_toplevel_mutation(runner: Runner, schema_name: &str, query: &str) -> TestResult<()> {
-        create_test_data(&runner, schema_name).await?;
-
-        match runner.query(query).await {
-            Ok(res) => res.assert_failure(2009, None),
-            Err(TestError::QueryConversionError(err)) if err.kind().code() == "P2009" => (),
-            Err(err) => return Err(err),
-        }
-
-        Ok(())
-    }
-
     async fn create_test_data(runner: &Runner, schema_name: &str) -> TestResult<()> {
         migrate_view(runner, schema_name).await?;
