+++ query-engine/connector-test-kit-rs/query-engine-tests/tests/writes/mod.rs
@@ -8,3 +8,4 @@ mod relations;
 mod top_level_mutations;
 mod unchecked_writes;
 mod uniques_and_node_selectors;
+mod views;
